+++ docs/make.jl
@@ -14,6 +14,7 @@ using MCMCChains: MCMCChains
 using Turing: Turing
 
 links = InterLinks(
+    "AdvancedHMC" => "https://turinglang.org/AdvancedHMC.jl/stable/",
     "DimensionalData" => "https://rafaqz.github.io/DimensionalData.jl/stable/",
     "DynamicPPL" => "https://turinglang.org/DynamicPPL.jl/stable/",
     "MCMCDiagnosticTools" => "https://turinglang.org/MCMCDiagnosticTools.jl/stable/",
@@ -66,4 +67,5 @@ if isnothing(old_GKSwstype)
     delete!(ENV, "GKSwstype")
 else
     ENV["GKSwstype"] = old_GKSwstype
-end;
+end
+nothing